refactor(migration): move migration resolve functions to util folder
